api/requirements.txt
amqp==5.1.1
asgiref==3.4.1
billiard==3.6.4.0
certifi==2021.10.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.5
djangorestframework==3.13.1
django-environ==0.8.1
django-timezone-field==4.2.3
kombu==5.2.4
pip==21.2.2
prompt-toolkit==3.0.29
psycopg2==2.8.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
setuptools==61.2.0
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wincertstore==0.2
django-filter